refactor (hickey + lowy): the boundary gets a name that says what it is, and the fence stops carrying a library
Both lenses over the diff, then my own reading against the evidence rather than
against either report. No `.agency/hickey.md` or `.agency/lowy.md` in this repo;
Lowy ran against `docs/architecture.md`'s own ladder and said so.

## `src/ui/` is `src/appliance/`, because the boundary is measurable

Two independent readers could not place a new file. The documented rule
contradicted itself inside one paragraph — the fold is argued as *"it is olai's
DRAWING of kolu"* and the rule then reads *"`src/ui/` is what KOLU draws"* — and
at the import site a reader got `ui` beside `browser`, two words each true of
both directories. The fold traded a name that said *kolu* for one that says *ui*.

So I measured what is actually behind each, and the answer is sharp enough to be
the rule:

    src/ui/       reaches @kolu/solid-dockrow, @kolu/solid-statepip,
                  @kolu/terminal-vocab, terminal-themes, @xterm/*,
                  @olai/kolu-client/wire            — eleven specifiers
    src/browser/  reaches NOTHING outside this package — zero

That is Löwy's naming test answered by the tree: one half stands on the
appliance's own product tier, the other on nothing but what olai hands across.
`src/appliance/` says it, `import … from "../appliance/index.ts"` inside
`browser/Feed.tsx` states the boundary at the call site, and it explains why odu
has no such directory and needs none — nothing odu draws reaches an odu
component tier, so there is no second half for a wall to hold. The rule is now
one sentence, from that reading, in the README, the manifest and
`docs/architecture.md`, which previously offered three formulations.

I did NOT add a test for it. A fence claim about a directory inside somebody's
package would be the fence inventing a layout convention and enforcing its own
invention — which is the argument the `.tsx` claim already makes about itself —
and the dangerous version (product tier outside the tenant) is held already.

## The fence stopped carrying a library

`fence.test.ts` was ~1430 lines, the first ~550 of them generic tree-reading:
glob expansion, manifest parsing, a source walk, four import grammars, a
transpiler, a resolver, a graph walk. Two concerns in one module, and a reader of
claim 6 had to scroll past a discriminated resolver to reach it. "Its only
consumer is this file" is circular — it lived there because it was written there.

`tree.testlib.ts` is the reading, and it knows nothing about plugins;
`fence.test.ts` is the nine claims and their reasons. The line is exactly that:
`namesAPlugin`, the tenants, the container and every claim stayed. 1430 → 1037 +
460, and `mechanics.test.ts` stopped carrying its own copy of the source walk —
the same recursion and the same `node_modules` argument, written twice in one
package with no wall between them, beside a one-level `readdirSync(PACKAGES)`
that the nesting had turned into a reading keying a tenant's files under
`plugins`. Sharing a READER is not sharing a CLAIM, and its header says so.

## One expansion of the workspace field, because two had already drifted

Three readers expand the root manifest's `workspaces` globs. Two are shell, and
in this very session they diverged: I gave `check-hydrated-deps.sh` a
matched-nothing refusal and left `prove-fence.sh` producing a short list in
silence — the exact failure both files carry paragraphs against.
`scripts/workspace-members.sh` is the one shell expansion; both read it.
`tree.testlib.ts` keeps its own, because a `bun test` corpus that shelled out
would make every equality in the fence depend on a subprocess — what is shared
is the RULE (refuse a shape you cannot expand, refuse a glob that matched
nothing), not the code.

## The testid collision is unrepresentable rather than checked

`{ ...appliance, ...CHROME }` could drop an id with no diagnostic. This has now
been wrong three ways and the third is the interesting one: prose first, then a
test that could not fail (recorded last pass), then a pair of module-scope
throws — and `tsc` told me the value throw was DEAD CODE, in as many words
("...the types '"terminal-block" | …' and '"padi" | …' have no overlap"). An
assertion the compiler can prove vacuous is one it can also prove. Both halves
are type-level assertions now: a collision is a `tsc` error naming the key or
the value, in every typecheck and every editor, with no runtime code and no test
to keep green. Proven by planting all four combinations.

One limit, measured rather than assumed after the first draft of that paragraph
claimed the opposite: the KEY assertion survives an `as const` being dropped; the
VALUE assertion does not (`Extract<string, "padi" | …>` is `never`, empty for the
wrong reason). Both words are load-bearing and the header now says which.

## ...and the layer the two inner ones left open

`selector(id: TestId | PluginTestId)` treats the app's table and the plugins'
merged table as ONE namespace of attribute values, and nothing compared them.
The fold is what made it bite: kolu's ten `terminal*`/`events*` ids used to sit
behind a package door nothing imported and are now spread into `PLUGIN_TESTID`,
reachable through that seam — the collision surface grew by ten and the check
for it did not exist. `packages/web/src/client/testids.test.ts` holds it, and
fires when planted.

Lowy's one action is taken: `docs/architecture.md` claimed `kolu-client` is "the
only package that names padi", which is true of the SOCKET and false of `@kolu/*`
— the tenant legitimately names six of them and the derived tenancy permits it.
Both spellings narrowed.

Named and not built, per each lens's own prove-then-extract rule: a leaf contract
package (a plugin cannot type-check its own manifest, because the package it
would import the type from imports it); TypeScript composite project references
as a compiler-enforced replacement for the package-granularity claims, which
would not subsume claims 4 and 5. Declined with reasons: splitting this PR (the
lane's brief is one PR, and the fence was rewritten in #447 — this adapts it);
and a test defending the appliance/browser filing line.
