+++ packages/plugin-api/README.md
@@ -2,7 +2,7 @@
 
 olai integrates with two things that are not olai. [kolu](https://kolu.dev) runs coding agents in terminals and serves them to an agent over MCP; [odu](https://github.com/juspay/odu) runs CI. Both were extracted into packages of their own once — [`@olai/kolu-client`](../kolu-client/README.md), `@olai/kolu-ui` and [`@olai/odu-client`](../odu-client/README.md) — and both left a residue behind in the packages that are supposed to know nothing about them: a `kolu.ts` in `@olai/chat`, a `koluConfig.ts` and a `claimants.ts` in `@olai/server`, a named `wiring.kolu` slot with a `koluHalf(…)` call beside it, four `...koluMembers` spreads in the middle of the wire spec, a row per plugin member in the server's expose map, a `padi/` folder in `@olai/web`, and one property key spelled at seven sites across four packages.
 
-That residue is not sloppiness. It is the part that genuinely was **olai's own judgement about an appliance** — what an absent padi means, which vault file is kolu's by convention, which property wears which face. What was missing is a place to put a judgement about an appliance that is neither the appliance nor the core. **This is that place, and it is the only one.** The tenants that stand on it — [`olai-plugin-kolu`](../plugins/olai-plugin-kolu/README.md) and [`olai-plugin-odu`](../plugins/olai-plugin-odu/README.md) — live one directory over, in `packages/plugins/`, and `@olai/kolu-ui` is not a package any more: the appliance fold made it `src/ui/` inside the kolu tenant, so an appliance is TWO packages (its dial, and everything olai says and draws about it) rather than three.
+That residue is not sloppiness. It is the part that genuinely was **olai's own judgement about an appliance** — what an absent padi means, which vault file is kolu's by convention, which property wears which face. What was missing is a place to put a judgement about an appliance that is neither the appliance nor the core. **This is that place, and it is the only one.** The tenants that stand on it — [`olai-plugin-kolu`](../plugins/olai-plugin-kolu/README.md) and [`olai-plugin-odu`](../plugins/olai-plugin-odu/README.md) — live one directory over, in `packages/plugins/`, and `@olai/kolu-ui` is not a package any more: the appliance fold made it `src/appliance/` inside the kolu tenant, so an appliance is TWO packages (its dial, and everything olai says and draws about it) rather than three.
 
 ## A plugin is a value
