+++ packages/plugins/olai-plugin-kolu/src/appliance/props/TerminalDoor.tsx
