+++ packages/plugins/olai-plugin-kolu/src/appliance/props/terminal.ts
