+++ packages/plugins/olai-plugin-kolu/src/index.ts
@@ -8,7 +8,7 @@
  * names no olai package at all, which the resolver proves and no sweep has to.
  * That is the sixth sitting's ruling and this package does not reopen it.
  *
- * What DID move is everything on this side of that wall. `./ui/` is EVERYTHING
+ * What DID move is everything on this side of that wall. `./appliance/` is EVERYTHING
  * KOLU RENDERS — the Dock row a terminal property wears, the live pane, the
  * re-attach policy, the fleet a tab holds once, and the words the header readout
  * says — and it was `@olai/kolu-ui` until the appliance fold. A second manifest
@@ -28,12 +28,12 @@
  * ## Which is why there are TWO browser directories and not one
  *
  * `./browser/` holds the padi PILL and the feed its press opens, and the tab's
- * own MOUNT. It is not a second `./ui/` and the line between them is exact:
- * `./ui/` draws what KOLU draws — its row, its emulator, its events — and this
+ * own MOUNT. It is not a second `./appliance/` and the line between them is exact:
+ * `./appliance/` draws what KOLU draws — its row, its emulator, its events — and this
  * draws what OLAI says about kolu in olai's own chrome. That line is why the
  * fold left two directories rather than merging them into one: it is a real
  * split and worth keeping visible, and what it never was is a PACKAGE split.
- * The pill's words are `./ui/`'s (`padiSaid`); the chip it sits in, the seat it takes in
+ * The pill's words are `./appliance/`'s (`padiSaid`); the chip it sits in, the seat it takes in
  * the bar and the drawer it opens are the app's judgement, and this is where
  * that judgement lives.
  *
@@ -49,7 +49,7 @@
  * `@olai/plugin-api` imports this package, so an import back would be a cycle the
  * manifests could not express. The fit is proved at the registry's
  * `satisfies`, which is the same structural agreement `@olai/ops` keeps with
- * the surface's `Status` and `./ui/`'s `props/block.ts` already keeps with the drawer's
+ * the surface's `Status` and `./appliance/`'s `props/block.ts` already keeps with the drawer's
  * entry — and it means a manifest that stopped fitting is named on the
  * registry's line, with this plugin's name on it.
  */
